src/services/capital_budgeting_service.py
from typing import List, Dict, Optional
import numpy as np
import numpy_financial as npf # We might need to add this dependency or implement manually if simple
from src.core.logging_config import LoggerMixin

class CapitalBudgetingService(LoggerMixin):
    """
    Service for Capital Budgeting and Fundamental Financial Modeling.
    Includes NPV, IRR, Payback Period, Profitability Index.
    """

    def calculate_npv(self, rate: float, cash_flows: List[float]) -> float:
        """
        Calculate Net Present Value.
        Args:
            rate: Discount rate (as decimal, e.g., 0.10 for 10%)
            cash_flows: List of cash flows starting from t=0 (initial investment usually negative)
        """
        try:
            # Manual implementation to avoid extra heavy dependencies if possible, 
            # but numpy_financial is standard. Let's start with pure numpy/python for simplicity 
            # or verify if numpy has it (numpy.npv was deprecated).
            # NPV = sum(Cf / (1+r)^t)
            values = np.array(cash_flows)
            times = np.arange(len(values))
            present_values = values / (1 + rate) ** times
            return float(np.sum(present_values))
        except Exception as e:
            self.logger.error(f"Error calculating NPV: {e}")
            raise

    def calculate_irr(self, cash_flows: List[float]) -> Optional[float]:
        """
        Calculate Internal Rate of Return.
        """
        try:
            # Using numpy's roots function to solve for IRR is robust enough for basic cases
            # or we can use the Newton-Raphson method.
            # For robustness, we will try to use numpy_financial if available, 
            # otherwise implement a simple solver.
            # Let's assume we'll add numpy-financial to requirements or implement a solver.
            # Implementing a simple trusted solver for now to keep deps low if desired, 
            # but usually IRR needs a good iterative solver.
            # We will use numpy implementation logic (finding roots of polynomial).
             
            # Roots of polynomial: C0 + C1*x + C2*x^2 ... where x = 1/(1+irr)
            # This is equivalent to finding roots of: C0*(1+r)^n + ... = 0
            
            # Simple approach: use numpy-financial logic (it's small)
            # But since I can't pip install easily without user permission on some envs, 
            # I'll stick to a basic root finding or ask to add numpy-financial.
            # I will add numpy-financial to requirements.txt later.
            import numpy_financial as npf
            return float(npf.irr(cash_flows))
        except ImportError:
            # Fallback if lib not present (though we should add it)
            self.logger.warning("numpy_financial not found, using simplified estimation")
            return 0.0 # Placeholder
        except Exception as e:
            self.logger.error(f"Error calculating IRR: {e}")
            raise

    def calculate_payback_period(self, cash_flows: List[float]) -> float:
        """
        Calculate Payback Period.
        """
        cumulative_cash_flow = np.cumsum(cash_flows)
        
        # If never positive, return infinity or error
        if cumulative_cash_flow[-1] < 0:
            return -1.0
            
        # Find first time it becomes positive
        # t where cumulative is positive
        positive_indices = np.where(cumulative_cash_flow >= 0)[0]
        if len(positive_indices) == 0:
            return -1.0
            
        t = positive_indices[0]
        
        if t == 0:
            return 0.0
            
        # Linear interpolation for more precision
        # Preceding cumulative value (negative)
        cumulative_prev = cumulative_cash_flow[t-1]
        cash_flow_t = cash_flows[t]
        
        fraction = abs(cumulative_prev) / cash_flow_t
        return float(t - 1 + fraction)

    def calculate_profitability_index(self, rate: float, cash_flows: List[float]) -> float:
        """
        Profitability Index = PV of Future Cash Flows / Initial Investment
        """
        if len(cash_flows) == 0:
            return 0.0
            
        initial_investment = abs(cash_flows[0])
        if initial_investment == 0:
            return 0.0
            
        future_cash_flows = cash_flows[1:]
        
        # Calculate PV of future flows
        values = np.array(future_cash_flows)
        # times starts at 1
        times = np.arange(1, len(values) + 1)
        present_values = values / (1 + rate) ** times
        pv_future = np.sum(present_values)
        
        return float(pv_future / initial_investment)
